Dependencies: Isolate `pyodbc` package
Per package metadata definition, it is an optional "extra" dependency.
Let's make it a reality, and also let CI cover the scenarios well.
