+++ Makefile
@@ -15,11 +15,13 @@ lock-deps:
 	@uv pip compile requirements.in --quiet -o requirements.txt 
 	@uv pip compile requirements.in --quiet -o requirements_arm64.txt --python-platform aarch64-unknown-linux-gnu
 
-deps: lock-deps
+deps: deps-full
+
+deps-minimal: lock-deps
 	uv pip install -r requirements-dev.txt
 
-deps-ci:
-	uv pip install --system -r requirements-dev.txt
+deps-full: lock-deps
+	uv pip install -r requirements-dev-full.txt
 
 test-ci:
 	set -a; source test.env; set +a; TESTCONTAINERS_RYUK_DISABLED=true pytest -n auto -x -rP -vv --tb=short --durations=10 --cov=ingestr --no-cov-on-fail
