+++ requirements-dev-full.txt
@@ -1,0 +1,2 @@
+-r requirements-dev.txt
+.[full]
