+++ requirements-dev.txt
@@ -6,7 +6,6 @@ pytest==8.3.3
 ruff==0.11.4
 hatchling==1.27.0
 build==1.2.1
-pyodbc==5.2.0
 twine==6.0.1
 testcontainers[postgres,mysql]==4.8.2
 pytest-xdist[psutil]==3.6.1
